Cleaned Up files and moved code into folder
