+++ app/src/main/java/com/example/campusguide/Bootstrapper.kt
@@ -8,7 +8,7 @@ import com.example.campusguide.location.FusedLocationProvider
 import com.example.campusguide.location.SwitchCampus
 import com.example.campusguide.map.GoogleMapAdapter
 import com.example.campusguide.map.GoogleMapInitializer
-import com.example.campusguide.map.ShowFloorPlan
+import com.example.campusguide.map.displayIndoor.ShowFloorPlan
 import com.example.campusguide.search.CustomSearch
 import com.example.campusguide.search.indoor.BuildingIndexSingleton
 import com.example.campusguide.search.indoor.IndoorLocationProvider
@@ -49,7 +49,8 @@ class Bootstrapper constructor(activity: MapsActivity) {
         activity.setOnCenterLocationListener(centerLocation)
 
         //Show Floor Plan
-        val showFloorPlan = ShowFloorPlan(map)
+        val showFloorPlan =
+            ShowFloorPlan(map)
         activity.setShowFloorPlanListener(showFloorPlan)
 
         // Switch Campus
